refactor:chat controller
